# backend-go/SessionService/docs/swagger.yaml
basePath: /
definitions:
...
  xxx_SessionService_models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Это пример API с gorilla/mux и swaggo
  title: Пример API
  version: "1.0"
paths:
...
  /session/{id}/nextQuestion:
    post:
      consumes:
      - application/json
      description: Advances to the next question in the session identified by the
        provided code.
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Successfully moved to the next question
        "405":
          description: Method not allowed
          schema:
            $ref: '#/definitions/xxx_SessionService_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/xxx_SessionService_models.ErrorResponse'
      summary: Move to the next question
      tags:
      - sessions
...
swagger: "2.0"